+++ src/main/java/fastcampus/board/service/HashtagService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -46,6 +47,10 @@ public class HashtagService {
         if (hashtag.getArticles().isEmpty()) {
             hashtagRepository.delete(hashtag);
         }
+    }
+
+    public List<String> getHashtags() {
+        return hashtagRepository.findAllHashtagNames();
     }
 
 }
